+++ src/meson.build
@@ -64,7 +64,6 @@ video_downloader_sources = files([
   'model.py',
   'playlist_dialog.py',
   'shortcuts_dialog.py',
-  'util.py',
   'window.py'
 ])
 python_sources_for_linting += video_downloader_sources
@@ -73,3 +72,4 @@ install_data(video_downloader_sources, install_dir: moduledir)
 
 
 subdir('downloader')
+subdir('util')
